fix border issues

+++ common/src/main/java/com/dfsek/terra/biome/pipeline/source/BiomeSource.java
@@ -3,5 +3,5 @@ package com.dfsek.terra.biome.pipeline.source;
 import com.dfsek.terra.api.world.biome.TerraBiome;
 
 public interface BiomeSource {
-    TerraBiome getBiome(int x, int z);
+    TerraBiome getBiome(double x, double z);
 }
